refactor: Move models, repos, middlewares, and google configuration into a single package
